fixes NC-2026 adds a `--network` option instead of separated options
Networks are now defined using a `--network` option that takes the cas insensitive
name of an enum values as input :   MAINNET, OTTOMAN, RINKEBY, ROPSTEN, GOERLI, DEV

These presets are a set of network-id, genesis file and boonodes list.

If the `--private-genesis-file` is provided with a valid JSON genesis file,
Pantheon consider that none of these predefined networks are used and requires
user to also set the `--network-id` and `--bootnodes` options.

However you can also just override the `--network-id` and `--bootnodes` for
a predefined network without having to provide a genesis file.