fixes NC-2026 adds a `--network` option instead of separated options
Networks are now defined using a `--network` option that takes the cas insensitive
name of an enum values as input :   MAINNET, RINKEBY, ROPSTEN, GOERLI, DEV

These presets are a set of network-id, genesis file and boonodes list.

If the `--genesis-file` is provided with a valid JSON genesis file,
Pantheon uses it instead of the default network. An empty bootnodes list is then
the default value and network id is the chain id found in the genesis file.

`--network-id` and `--bootnodes` options can override these defaults.

You can of course also override the `--network-id` and `--bootnodes` for a
predefined known network (using `--network`).

Also fixes NC-2189 renaming --private-genesis-file to --genesis-file

Updated a lot of doc according to the options changes